% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\docType{data}
\name{admin}
\alias{admin}
\title{Admin lookup table}
\format{An object of class \code{character} of length 78.}
\usage{
data(admin)
}
\description{
This is a lookup table for Problemcodes pertaining to administrative requests made to the CSB.
}
\keyword{datasets}
