% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{cat}
\alias{cat}
\alias{cat_admin}
\alias{cat_animal}
\alias{cat_construction}
\alias{cat_debris}
\alias{cat_degrade}
\alias{cat_disturbance}
\alias{cat_event}
\alias{cat_health}
\alias{cat_landscape}
\alias{cat_law}
\alias{cat_maintenance}
\alias{cat_nature}
\alias{cat_road}
\alias{cat_sewer}
\alias{cat_traffic}
\alias{cat_vacant}
\alias{cat_waste}
\title{Lookup Tables Used for Categorization of Problem Codes}
\format{An object of class \code{character} of length 78.}
\usage{
cat_admin

cat_animal

cat_construction

cat_debris

cat_degrade

cat_disturbance

cat_event

cat_health

cat_landscape

cat_law

cat_maintenance

cat_nature

cat_road

cat_sewer

cat_traffic

cat_vacant

cat_waste
}
\description{
Any data object from the stlcsb package that starts with `cat_` is a lookup table. These tables were hand-made,contact (foxbl@slu.edu) with questions or concerns. It is used internally in any function that categorizes by problem code. It is available to the user as a matter of convenience and transparency.
}
\keyword{datasets}
