% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_animal.R
\docType{data}
\name{cat_animal}
\alias{cat_animal}
\title{Animal lookup table}
\format{An object of class \code{character} of length 34.}
\usage{
cat_animal should be supplied as an argument for the csb_filter function.
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table, This is a lookup table for Problem codes pertaining to animal requests made to the CSB.
}
\keyword{datasets}
