% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_construction.R
\docType{data}
\name{cat_construction}
\alias{cat_construction}
\title{Construction lookup table}
\format{An object of class \code{character} of length 24.}
\usage{
cat_construction should be supplied as an argument for the \code{csb_filter} function.
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table, This is a lookup table for Problem codes pertaining to construction requests made to the CSB.
}
\keyword{datasets}
