% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_degrade.R
\docType{data}
\name{cat_degrade}
\alias{cat_degrade}
\title{Degrade lookup table}
\format{An object of class \code{character} of length 30.}
\usage{
cat_degrade should be supplied as an argument for the \code{csb_filter} function.
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table, This is a lookup table for Problem codes pertaining to the degrading of buildings or city property.
}
\keyword{datasets}
