% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_maintenance.R
\docType{data}
\name{cat_maintenance}
\alias{cat_maintenance}
\title{Maintenance lookup table}
\format{An object of class \code{character} of length 65.}
\usage{
maintenance should be supplied as an argument for the \code{csb_filter} function.
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table, This is a lookup table for Problem codes pertaining to internal maintenance requests made to the CSB.
}
\keyword{datasets}
