% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_nature.R
\docType{data}
\name{cat_nature}
\alias{cat_nature}
\title{Nature lookup table}
\format{An object of class \code{character} of length 9.}
\usage{
nature should be supplied as an argument for the \code{csb_filter} function.
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table, This is a lookup table for Problem codes pertaining to natural event (snow, etc) requests made to the CSB.
}
\keyword{datasets}
