% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_road.R
\docType{data}
\name{cat_road}
\alias{cat_road}
\title{Road lookup table}
\format{An object of class \code{character} of length 34.}
\usage{
road should be supplied as an argument for the \code{csb_filter} function.
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table, This is a lookup table for Problem codes pertaining to road requests made to the CSB.
}
\keyword{datasets}
