% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_vacant.R
\docType{data}
\name{cat_vacant}
\alias{cat_vacant}
\title{Vacant lookup table}
\format{An object of class \code{character} of length 19.}
\usage{
cat_vacant should be supplied as an argument for the \code{csb_filter} function.
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table, This is a lookup table for Problem codes pertaining to requests that may indicate vacant parcels. These problem codes coincide with the base categories, and this is the lookup table used in the csb_vacant function.
}
\keyword{datasets}
