% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_waste.R
\docType{data}
\name{cat_waste}
\alias{cat_waste}
\title{Waste lookup table}
\format{An object of class \code{character} of length 60.}
\usage{
waste should be supplied as an argument for the \code{csb_filter} function.
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table, This is a lookup table for Problem codes pertaining to waste (Trash and recycling) requests made to the CSB.
}
\keyword{datasets}
