% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_address.R
\name{csb_address}
\alias{csb_address}
\title{Clean CSB addresses}
\usage{
csb_address(.data, var, newVar)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{var}{The variable in the data that contains}

\item{newVar}{Name of output variable with a more tidy version of address.}
}
\value{
\code{csb_address} returns data with an additional variable for a more tidy address.

need to import postmastr!!
}
\description{
\code{csb_address} parses addresses in a manner better suited for geocoding.
}
