% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_api_key.R
\name{csb_api_key}
\alias{csb_api_key}
\title{Store CSB API Key}
\usage{
csb_api_key(key, install, overwrite)
}
\arguments{
\item{key}{The API key provided to you by the City of Saint Louis, formatted in quotes. A key can be aquired \href{https://www.stlouis-mo.gov/government/departments/information-technology/web-development/city-api/sign-up.cfm}{Here}}

\item{overwrite}{If TRUE, will overwrite an existing CSB_API_KEY that you already have in your \code{.Renviron}}

\item{install}{If TRUE, will install the key in your \code{.Renviron} file for use in future sessions. Defaults to FALSE.}
}
\description{
\code{csb_api_key} can be used to store your CSB API key in your \code{.Renviron} for repeated use.
Your API key will be securely called without having to reveal it in your code. After installing your key, it can be called any time by typing \code{Sys.getenv("CSB_API_KEY")} and can be
used in package functions by simply typing CSB_API_KEY If you do not have an \code{.Renviron} file, the function will create one for you.
If you already have an \code{.Renviron} file, the function will append the key to your existing file, while making a backup of your
original file for data recovery purposes.
  This function is adapted from the \href{https://github.com/walkerke/tidycensus}{tidycensus} package.
}
