% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_cancelled.R
\name{csb_cancelled}
\alias{csb_cancelled}
\title{Remove Cancelled Calls for Service}
\usage{
csb_cancelled(.data, var, drop = TRUE)
}
\arguments{
\item{.data}{A tbl}

\item{var}{Name of the column containing cancellation timestamps; by default
this should be \code{DATECANCELLED}}

\item{drop}{Optionally remove the now empty column that had contained cancellation
date and time if \code{TRUE}}
}
\value{
Returns a tibble with the rows containing dates and times for the
    given variable removed.
}
\description{
Subsets data to remove any call with a date and time cancelled.
}
\examples{
csb_cancelled(january_2018, var = "DATECANCELLED")
csb_cancelled(january_2018, var = "DATECANCELLED", drop = FALSE)

}
