% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_date.R
\name{csb_date}
\alias{csb_date}
\title{Clean CSB date}
\usage{
csb_date(.data, var, day = NULL, month = NULL, year = NULL, delete = FALSE)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{var}{name of column containing date data}

\item{day}{if specified, returns a named column with parsed day}

\item{month}{if specified, returns a named column with parsed month}

\item{year}{if specified, returns a named column with parsed year}

\item{filter}{if true, the function will filter data based on specified values rather than parse it}

\item{delete}{if true, deletes the original column with unparsed data}
}
\value{
\code{csb_date} can either return new columns with parsed date or return a filtered dataset
}
\description{
\code{csb_date} returns parsed dates in specified columns
}
