% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_date_parse.R
\name{csb_date_parse}
\alias{csb_date_parse}
\title{Parse CSB date}
\usage{
csb_date_parse(.data, var, day, month, year, delete = FALSE)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{var}{name of column containing date data}

\item{day}{if specified, returns a named column with parsed day}

\item{month}{if specified, returns a named column with parsed month}

\item{year}{if specified, returns a named column with parsed year}

\item{delete}{if true, deletes the original column with unparsed data}
}
\value{
\code{csb_date_parse} returns new columns with specified names containing parsed date information
}
\description{
\code{csb_date_parse} is used to parse out dates into more intellgible formats
}
\examples{
csb_date_parse(january_2018, DATETIMEINIT, dayInit)
csb_date_parse(january_2018, DATETIMEINIT, dayInit, monthInit)
csb_date_parse(january_2018, DATETIMEINIT, month = monthInit)
csb_date_parse(january_2018, DATETIMEINIT, month = monthInit, year = yearInit)
csb_date_parse(january_2018, DATETIMEINIT, dayInit, monthInit, yearInit, delete = TRUE)

}
