% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_geo.R
\name{csb_geo}
\alias{csb_geo}
\title{Clean CSB geo to SF object}
\usage{
csb_geo(.data, varX, varY, replace, crs)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{varX}{name of column containing SRX data}

\item{varY}{name of column containing SRY data}

\item{replace}{If true, the original spatial data columns will be dropped}

\item{crs}{coordinate reference system for the data to be projected into}
}
\value{
\code{csb_geo} returns a `latitude` and `longitude` column with properly formatted spatial data in an SF format
}
\description{
\code{csb_geo} converts SRX and SRY data into a simple features object.
}
