% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_get_data.R
\name{csb_get_data}
\alias{csb_get_data}
\title{Download CSB Data}
\usage{
csb_get_data()
}
\value{
\code{csb_get_data} returns a tibble with all of the CSB requests
    contained in our version of the CSB's data release.
}
\description{
\code{csb_get_data} provides direct access to a compiled version of the CSB's data release.
}
