% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_intersection.R
\name{csb_intersection}
\alias{csb_intersection}
\title{CSB intersections}
\usage{
csb_intersection(.data, var, newVar, filter = FALSE, remove = FALSE)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{var}{name of column containing address}

\item{newVar}{Name of output variable to be created with logical}

\item{filter}{If true, returns a filtered version of the input with only intersections}

\item{remove}{If true, removes intersections from the input data frame}
}
\value{
\code{csb_intersection}returns a logical vector for addresses that are intersections, or it filters intersections
}
\description{
\code{csb_intersection}returns a logical vector with TRUE for addresses that are intersections, or filters or drops these observations
}
