% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_last_update.R
\name{csb_last_update}
\alias{csb_last_update}
\title{Date of Last CSB Data Update from the City of St. Louis}
\usage{
csb_last_update()
}
\value{
A string scalar containing the date of last update.
}
\description{
Data are updated by the City of St. Louis on their
    \href{https://www.stlouis-mo.gov/data/}{open data} site on a
    weekly basis. This function returns the date of the last
    update.
}
\examples{
\dontrun{
last_update <- csb_last_update()
}
}
