% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_projectXY.R
\name{csb_projectXY}
\alias{csb_projectXY}
\title{Clean CSB geo to SF object}
\usage{
csb_projectXY(.data, varX, varY, crs)
}
\arguments{
\item{.data}{A tbl}

\item{varX}{Name of column containing SRX data}

\item{varY}{Name of column containing SRY data}

\item{crs}{Optional; coordinate reference system for the data to be projected into}
}
\value{
\code{csb_projectXY} returns a sf object of the input data, specifying a new crs will
    reproject the data.
}
\description{
\code{csb_projectXY} converts SRX and SRY data into a simple features object.
    You can write a shapefile directly from the output of this function using \code{sf::st_write}.
}
\examples{
#You MUST remove observations with missing coordinates before using this function
csb <- csb_missingXY(january_2018, SRX, SRX, newVar = missing)
csb <- dplyr::filter(csb, missing == FALSE)
csb_projectXY(csb, SRX, SRY)
csb_projectXY(csb, SRX, SRY, crs = 4269)

}
