% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_vacant.R
\name{csb_vacant}
\alias{csb_vacant}
\title{CSB vacant codes}
\usage{
csb_vacant(.data, var, newVar, filter = FALSE)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{var}{name of column containing problem codes}

\item{newVar}{name of output variable to be created with vacant logical}

\item{filter}{If true, returns a filtered version of the input with only vacancy related problem codes}
}
\value{
\code{csb_vacant}returns a logical vector with TRUE for vacancy related problem codes, or a filtered version of the input data
}
\description{
\code{csb_vacant}appends a logical vector indicating `TRUE` for vacancy related problem codes.
}
\examples{
csb_vacant(january_2018, PROBLEMCODE, vacant)
csb_vacant(january_2018, PROBLEMCODE, vacant, filter = TRUE)

}
