% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{cat}
\alias{cat}
\alias{cat_admin}
\alias{cat_animal}
\alias{cat_construction}
\alias{cat_debris}
\alias{cat_degrade}
\alias{cat_disturbance}
\alias{cat_event}
\alias{cat_health}
\alias{cat_landscape}
\alias{cat_law}
\alias{cat_maintenance}
\alias{cat_nature}
\alias{cat_road}
\alias{cat_sewer}
\alias{cat_traffic}
\alias{cat_vacant}
\alias{cat_waste}
\title{Lookup Tables Used for Categorization of Problem Codes}
\format{An object of class \code{character} of length 78.}
\usage{
cat_admin

cat_animal

cat_construction

cat_debris

cat_degrade

cat_disturbance

cat_event

cat_health

cat_landscape

cat_law

cat_maintenance

cat_nature

cat_road

cat_sewer

cat_traffic

cat_vacant

cat_waste
}
\value{
A named character vector listing the specific problem codes associated with
    each category.
}
\description{
Any data object from the \code{stlcsb} package that starts with \code{cat_}
    is a lookup table. These tables were hand-made - contact the package
    maintainers with questions or concerns. These tables are used internally
    in any function that categorizes by problem code. They are available to the user
    as a matter of convenience and transparency.
}
\keyword{datasets}
