% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_categories.R
\docType{data}
\name{cat_law}
\alias{cat_law}
\title{Lookup Tables Used for Categorization of Problem Codes}
\format{An object of class \code{character} of length 21.}
\usage{
cat_law
}
\description{
Any data object from the stlcsb package that starts with \code{cat_} is a lookup table. These tables were hand-made,contact (foxbl@slu.edu) with questions or concerns. It is used internally in any function that categorizes by problem code. It is available to the user as a matter of convenience and transparency.
}
\keyword{datasets}
