% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_cancelled.R
\name{csb_cancelled}
\alias{csb_cancelled}
\title{Remove Cancelled Calls for Service}
\usage{
csb_cancelled(.data, var, drop = TRUE)
}
\arguments{
\item{.data}{A tbl or data frame}

\item{var}{Name of the column containing cancellation timestamps}

\item{drop}{A logical sclar; if \code{TRUE}, removes the now empty column
that had contained cancellation date and time, otherwise if \code{FALSE}
the empty column is retained.}
}
\value{
Returns a tibble with the rows containing dates and times for the
    given variable removed.
}
\description{
Subsets data to remove any call with a date and time cancelled.
}
\examples{
csb_cancelled(january_2018, var = "datecancelled")
csb_cancelled(january_2018, var = "datecancelled", drop = FALSE)

}
