% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_categorize.R
\name{csb_categorize}
\alias{csb_categorize}
\title{Categorize Raw CSB Data}
\usage{
csb_categorize(.data, var, newVar)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{var}{Name of orginial variable containing problem code}

\item{newVar}{Name of output variable to be created with category name}
}
\value{
\code{csb_categorize} returns data with an additional variable for an intelligible category for CSB requests.
}
\description{
\code{csb_categorize} provides intelligible categories for the CSB data based on problem code.
}
