% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_categorize.R
\name{csb_categorize}
\alias{csb_categorize}
\title{Categorize CSB Call Types}
\usage{
csb_categorize(.data, var, newVar)
}
\arguments{
\item{.data}{A tibble or data frame}

\item{var}{Name of existing column containing problem codes}

\item{newVar}{Name of output variable to be created with category string}
}
\value{
Returns a tibble with the string vector added as a new variable.
}
\description{
\code{csb_categorize} provides general categories for the
    CSB data based on problem code. These were created based on a review
    of the call data in mid-2018.
}
\examples{
csb_categorize(january_2018, var = problemcode, newVar = Category)

}
