% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_date_filter.R
\name{csb_date_filter}
\alias{csb_date_filter}
\title{Filter CSB date}
\usage{
csb_date_filter(.data, var, day, month, year, delete = FALSE)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{var}{name of column containing date data}

\item{day}{numeric representation of day(s) to include (leading zero or not will be accepted)}

\item{month}{numeric/character representation of month(s) to include}

\item{year}{numeric representation of years(s) to include (4 digit or 2 digit accepted)}

\item{delete}{if true, deletes the original column containing dates}
}
\value{
\code{csb_date_filter}returns a filtered version of the input data based on specified arguments
}
\description{
\code{csb_date_filter}Filters dates to return only the specified date elements
}
