% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_geo.R
\name{csb_geo}
\alias{csb_geo}
\title{Clean CSB geo to SF object}
\usage{
csb_geo(.data, varX, varY, crs = NULL)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{varX}{name of column containing SRX data}

\item{varY}{name of column containing SRY data}

\item{crs}{coordinate reference system for the data to be projected into}
}
\value{
\code{csb_geo} returns a sf object of the input data, specifying a new crs will reproject the data.
}
\description{
\code{csb_geo} converts SRX and SRY data into a simple features object. You can write a shapefile directly from the output of this function using sf::st_write
}
\examples{
#You MUST remove observations with missing coordinates before using this function
csb <- csb_missing(january_2018, SRX, SRX, filter = TRUE)
csb_geo(csb, SRX, SRY)
csb_geo(csb, SRX, SRY, crs = 4269)

}
