% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_intersection.R
\name{csb_intersection}
\alias{csb_intersection}
\title{Identify Calls for Services at Street Intersections}
\usage{
csb_intersection(.data, var, newVar, filter = FALSE, remove = FALSE)
}
\arguments{
\item{.data}{A tbl}

\item{var}{name of column containing address}

\item{newVar}{Name of output variable to be created with logical}

\item{filter}{If true, returns a filtered version of the input with only observations occurring at intersections}

\item{remove}{If true, returns a version of the input without observations occuring at intersections}
}
\value{
\code{csb_intersection}returns a logical vector for addresses that are intersections, or it filters intersections
}
\description{
\code{csb_intersection}returns a logical vector indicating `TRUE` for addresses that are intersections. This function is also able to isolate or remove these data.
}
\examples{
\dontrun{csb_intersection(january_2018, PROBADDRESS, newVar = Intersection)}
csb_intersection(january_2018, PROBADDRESS, filter = TRUE)
csb_intersection(january_2018, PROBADDRESS, remove = TRUE)

}
