% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_missing.R
\name{csb_missing}
\alias{csb_missing}
\title{Missing CSB data}
\usage{
csb_missing(.data, varX, varY, newVar, filter = FALSE)
}
\arguments{
\item{.data}{A tibble with raw CSB data}

\item{varX}{name of column containing SRX data}

\item{varY}{name of column containing  SRY data}

\item{newVar}{if named, produces a new column with a logical indicator of incomplete spatial data}

\item{filter}{if true, returns a filtered version of the data with only observations with complete spatial data}
}
\value{
\code{csb_missing} returns a logical vector indicating `TRUE` if an observation has incomplete spatial data, or a filtered version of the input data
}
\description{
\code{csb_missing} returns a logical vector indicating `TRUE` if an observation is missing SRX or SRY data.
}
