% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_missingXY.R
\name{csb_missingXY}
\alias{csb_missingXY}
\title{Missing spatial CSB data}
\usage{
csb_missingXY(.data, varX, varY, newVar)
}
\arguments{
\item{.data}{A tbl}

\item{varX}{name of column containing SRX data}

\item{varY}{name of column containing SRY data}

\item{newVar}{name of new column that is \code{TRUE} if coordinate data are
missing and \code{FALSE} otherwise.}
}
\value{
\code{csb_missingXY} A tbl with a logical vector appended to it.
}
\description{
\code{csb_missingXY} returns a logical vector indicating if an
    observation is missing in either of the coordinate columns that come
    with the CSB data.
}
\examples{
csb_missingXY(january_2018, SRX, SRY, newVar = "missingXY")

}
