% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csb_missingXY.R
\name{csb_missingXY}
\alias{csb_missingXY}
\title{Identifying Calls Missing Coordinate Data}
\usage{
csb_missingXY(.data, varX, varY, newVar)
}
\arguments{
\item{.data}{A tibble or data frame}

\item{varX}{Name of column containing x coordinate data}

\item{varY}{Name of column containing y coordinate data}

\item{newVar}{Name of new column that is \code{TRUE} if coordinate data are
missing and \code{FALSE} otherwise.}
}
\value{
A tbl with a logical vector appended to it.
}
\description{
\code{csb_missingXY} returns a logical vector indicating if an
    observation is missing in either of the coordinate columns that come
    with the CSB data.
}
\examples{
csb_missingXY(january_2018, srx, sry, newVar = "missingXY")

}
